# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Vinsky API Documentation'
  description: ''
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
...
  /api/v1/me/password:
    put:
      summary: 'Update password'
      operationId: updatePassword
      description: ''
      parameters: []
      responses: {}
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: ''
                  example: securePassword123
                password:
                  type: string
                  description: 'Must be at least 8 characters.'
                  example: newSecurePassword123
              required:
                - current_password
                - password